requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==2.2.7
django-cors-headers==3.2.0
django-extensions==2.2.5
djangorestframework==3.10.3
ffmpeg==1.4
ffprobe==0.5
gTTS==2.2.1
gTTS-token==1.1.4
idna==2.8
mysqlclient==1.4.6
nltk==3.4.5
pydot==1.4.1
pydub==0.23.1
PyMySQL==0.9.3
pyparsing==2.4.5
pytz==2019.3
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
sqlparse==0.3.0
urllib3==1.25.7
Wave==0.0.2
